+++ utils/helpers.go
@@ -1,8 +1,9 @@
 package utils
 
 import (
-	"github.com/blutspende/bloodlab-common/timezone"
 	"time"
+
+	"github.com/blutspende/bloodlab-common/timezone"
 )
 
 func FormatTimeStringToBerlinTime(timeString, format string) time.Time {
@@ -15,4 +16,16 @@ func FormatTimeStringToBerlinTime(timeString, format string) time.Time {
 		return time.Time{}
 	}
 	return result
+}
+
+func ParseBerlinTimeStringToUTCTime(timeString string) time.Time {
+	location, err := timezone.EuropeBerlin.GetLocation()
+	if err != nil {
+		return time.Time{}
+	}
+	berlinTime, err := time.ParseInLocation("20060102150405", timeString, location)
+	if err != nil {
+		return time.Time{}
+	}
+	return berlinTime.UTC()
 }