+++ CHANGELOG.md
@@ -5,6 +5,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [1.0.3] - 2026-01-16
+
+### Added
+- Berlin time string to UTC time helper
+
+### Changed
+
+### Fixed
+
 ## [1.0.2] - 2025-07-18
 
 ### Added